clean & refactor
