+++ src/main/java/org/scm/ojt/rest/entity/BaseEntity.java
@@ -1,27 +1,26 @@
-package org.scm.ojt.rest.model;
+package org.scm.ojt.rest.entity;
 
 import lombok.Data;
-import lombok.Getter;
 import org.bson.types.ObjectId;
 import org.mongodb.morphia.annotations.Id;
 import org.mongodb.morphia.annotations.Property;
 import org.mongodb.morphia.annotations.Version;
 
 /**
- * Created by Wicaksono on 10/11/2018.
+ * Created by Yoyok_T on 28/09/2018.
  */
 @Data
-public abstract class BaseCollection {
+public abstract class BaseEntity {
     @Id
     @Property("id")
-    private ObjectId id;
+    protected ObjectId id;
 
     @Version
     @Property("version")
     private Long version;
 
-//    public String getID(){
-//        return getID().toString();
-//    }
+    public BaseEntity() {
+        super();
+    }
 
 }
