+++ src/main/java/org/scm/ojt/rest/entity/EntityModel.java
@@ -1,4 +1,4 @@
-package org.scm.ojt.rest.model;
+package org.scm.ojt.rest.entity;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
