+++ src/main/java/org/scm/ojt/rest/filter/CrossOriginResourceSharingFilter.java
@@ -5,9 +5,6 @@ import java.io.IOException;
 import javax.servlet.*;
 import javax.servlet.http.HttpServletResponse;
 
-/**
- * @author cluttered.code@gmail.com
- */
 public class CrossOriginResourceSharingFilter implements javax.servlet.Filter {
 
     @Override
