+++ src/main/java/org/scm/ojt/rest/logic/TestLogic.java
@@ -1,11 +1,8 @@
 package org.scm.ojt.rest.logic;
 
 import org.scm.ojt.rest.dao.TestDAO;
-import org.scm.ojt.rest.model.Employee;
-import org.mongodb.morphia.Datastore;
 
 import javax.inject.Inject;
-import javax.ws.rs.core.Response;
 
 public class TestLogic {
